requirements.txt
psycopg2-binary>=2.9.11
sqlalchemy>=2.0.46
python-dotenv>=1.2.1
structlog>=25.5.0
line_profiler>=5.0.1
pika>=1.3.2
requests>=2.32.5
bs4>=0.0.2
lxml>=6.0.2
python-dateutil>=2.9.0
pandas>=3.0.0
html5lib>=1.1
pydantic>=2.12.5
pydantic-settings>=2.12.0
cryptography>=46.0.5
httpx>=0.28.1
zeep>=4.3.2
tldextract>=5.1.0
croniter>=1.4.1
dramatiq[rabbitmq]>=2.1.0
watchdog>=6.0.0
pywin32>=310; sys_platform == "win32"
pdfplumber>=0.11.0
pytesseract>=0.3.13
Pillow>=11.0.0
opencv-python-headless>=4.10.0
paddleocr>=3.0.0
paddlepaddle>=3.0.0
